Made Changes to READ and WRITE functions
Changes (Not Tested fully):
Within the WRITE function, added a check to see whether the data would overflow to next page or to the next sector, if data overflows to the next page, we can do a recursive call to the WRITE function to write the overflown data to memory with appropriate measures.  If overflown to next sector, we simply set clearFlag variable to 1 so whenever the WRITE function is called the next time, it will clear the memory before writing.

Bug Fixes:
Fixed the bit shifting for circular buffer head and tail that could have caused trouble in the future

Instead of Writing 100 bytes to the memory module, packet size is calculated and only the data is written to the memory fixing the issue of overwriting previous data

When Reading from Memory, changed the buffer size to the size of the last data that was written so only that is received into the buffer when reading

Potential Remaining Bugs:
MEM_CLEAR function needs to be called twice because ENABLE_WREN function doesn't set the WEL latch to 1 100% of the times. This seems to only happen when calling MEM_CLEAR function. Any other time when ENABLE_WREN is called, it seems to work as far as tested.